Edited the model-package to use doubles instead of floats. Our original idea to adapt our model to JME is flawed since we should not adapt to JME. It should be adapted to our model instead. I.E typecasting will be done later when connecting JME with our model.

+++ project/src/Model/Battlefield.java
@@ -25,7 +25,7 @@ public class Battlefield {
      * @param xWidth 
      * @param yWidth 
      */
-    public Battlefield(float xWidth, float yWidth) {
+    public Battlefield(double xWidth, double yWidth) {
         // Checking for errors
         if(xWidth == 0 && yWidth != 0) {
             xWidth = yWidth;
